Refactor objectStore to destructibleStore
