+++ three-fps/src/Game/Terrain/Cube.tsx
@@ -3,14 +3,14 @@ import { useLoader } from "@react-three/fiber"
 import { useBox } from "@react-three/cannon"
 import dirt from "../Assets/dirt.jpg"
 import { useSelector } from "react-redux"
-import { RootState } from "../Reducers/objectStore"
+import { RootState } from "../Reducers/GameStore"
 
 /**
  * Simple Cube Mesh that can take damage
- * 
- * adapted in part from Maksim Ivanov 
+ *
+ * adapted in part from Maksim Ivanov
  * "React Minecraft": https://www.youtube.com/watch?v=Lc2JvBXMesY&t=124s
- * @param {*} props 
+ * @param {*} props
  * @returns A Cube Mesh
  */
 const CubeMesh = (props : any) => {
@@ -27,7 +27,7 @@ const CubeMesh = (props : any) => {
 }
 
 export function Cube(props : any) {
-	const health = useSelector((state : RootState) => state.objects[props.objectID].health)
+	const health = useSelector((state : RootState) => state.destructibles.destructibles[props.objectID].health)
 
 	if (health > 0) {
 		return (
